Same functionality, small refactor.
