+++ src/main/java/com/diffplug/blowdryer/Blowdryer.java
@@ -294,4 +294,38 @@ public class Blowdryer {
 			return (T) value;
 		}
 	}
+
+	/** Alias for {@link Blowdryer} which fills in the `project` field automatically. */
+	public static class WithProject {
+		private final Project project;
+
+		public WithProject(Project project) {
+			this.project = project;
+		}
+
+		/** Alias for {@link Blowdryer#immutableUrl(String)}. */
+		public File immutableUrl(String url) {
+			return Blowdryer.immutableUrl(url);
+		}
+
+		/** Alias for {@link Blowdryer#file(String)}. */
+		public File file(String resource) {
+			return Blowdryer.file(resource);
+		}
+
+		/** Alias for {@link Blowdryer#prop(String, String)}. */
+		public String prop(String propFile, String key) throws IOException {
+			return Blowdryer.prop(propFile, key);
+		}
+
+		/** Alias for {@link Blowdryer#cfg(Project, String, String)}. */
+		public String cfg(String key, String descForError) {
+			return Blowdryer.cfg(project, key, descForError);
+		}
+
+		/** Alias for {@link Blowdryer#cfg(Project, Class, String, String)}. */
+		public <T> T cfg(Class<T> clazz, String key, String descForError) {
+			return Blowdryer.cfg(project, clazz, key, descForError);
+		}
+	}
 }
